src/components/servo/dom/bindings/codegen/test/test_bug788369.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=788369
-->
<head>
  <meta charset="utf-8">
  <title>Test for Bug 788369</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=788369">Mozilla Bug 788369</a>
<p id="display"></p>
<div id="content" style="display: none">
</div>
<pre id="test">
<script type="application/javascript">

/** Test for Bug 788369 **/
try {
  var xhr = new(window.ActiveXObject || XMLHttpRequest)("Microsoft.XMLHTTP");
  ok(xhr instanceof XMLHttpRequest, "Should have an XHR object");
} catch (e) {
  ok(false, "Should not throw exception when constructing: " + e);
}
</script>
</pre>
</body>
</html>
